main: Update the seek slider hack for GTK+ changes
GTK+ commit 162614aa changed the range widget's button behaviour so that
button 1 jumps to position. This is almost what we need in totem, except
that we'd want buttons 2 and 3 to also do the jumping.

Update the seek slider hack accordingly to always force button 1
behaviour and bump gtk+ dep to 3.5.2.

https://bugzilla.gnome.org/show_bug.cgi?id=675838